address issue #3426: maintain a table of unresolved references
These unresolved references are local links of the form [[...]]). Since
backlink references (e.g. {{my-references}}) are mainted on on page edits
via ids, dead links could not be maintained this way.

A new table is created to maintain such dead links, and refreshes references
to dead links, when such pages are created.