fix bug: never update references in the search renderer
before, it was possible that the search renderer added some
unsresolved references, although these were already fixed.
This was caused by the asynchrnous search queue, which
might have inserted such values.

The interface of Page.render was extended to be able to
specify, that the references should be always updated in the DB,
or just when unresolved references are encountered, or
never.