generalize handling of langstrings
this change updates yesterday's change with ml_title. It generalizes
the code by introducing an instproc named "update_langstring_property" that
generalizes setting of arbitrary "properties" based on a property "langstring"
that contains a dict of attributes in multiple languages (which are part
of the page, make it suitable also when the enduser cannot enter
message keys).