added new form-field type: comp_correct_when added conjunctive answer-check adde answer-check "contains-not"
The new form-field type offers a graphical user interface for input of
correct when clauses. When comp_correct_when is used with the repeat
flag it produces conjunctive (AND connected) clauses. The answer-check
"contains-not" is more or less the same a "contains", but highlights
such words in red.

bump version number to 5.10.0d71