More restricted checking for object_ids
This avoids SQL errors in cases, where values are passed in via query variables
which are valid integers but out of range for SQL integer, which are 32-bit
integers. The better solution would even be to define an NSF argument type
"object_id", but unfortunately, this requires a newer version of NSF than
we target for the release.