Added archiving of items, reduced usage of images for icons
When an item is archived, it is not shown to plain users. Internally,
when an item is archived, the publish status is changed to "expired"
and removed from search/syndication.  When an item is archived, it can
be brought via "toggle-publish-status" to state "production" again.

Future versions might rename the item in such conditions, but
currently, it is NOT renamed, so one cannot create one more item in
the same folder with the same name.

For realizing this functionality, the includelet "form-usages" accepts
now an additional optional value for "-buttons" named "archive". The
publish status toggle can be used now for switching from
"production"->"ready", "ready"->"production" and
"expired"->"production".

Additionally, the usage of .png images as icons was reduced by CSS
elements for easier styling (e.g., for different themes) and for
reducing less server requests and better responsiveness. In particular
the images for active.png and inactive.png are not necessary any more
and were removed.

Bumped version number to 5.10.1d8 to enforce loading of the message
keys.