Address issue with .SELF. references in exams (xowf)
Since it is the goal of the exams to store all resources directly
under the questions, handling extra .SELF. references (which are
inserted by a downstream plugin for CKeditor) do not make sense. A
probably cleaner solution to this is to configure the CKeditor
to omit the .SELF. references in these cases.