Support specification of allowed tags/attributes/protocols via global package parameters.
This change makes it possible to define for a package values for
AllowedTags, AllowedAttributes, or AllowedProtocols.  If (some of)
these exist, use these for configuring "ad_html_security_check". If +
not, fall back to the default (site wide) definition.

This change is fully backward compatible and effects on sites defining
such global parameters. This change was induced by the need of large
sites, where it is not feasible to set these parameters on the package
instance level, since this site has 830000 packages instances.

Bump version number to 5.10.1d19