Extend randomization of multiple alternatives for cases, where one alternative must be "true"
Certain evaluation schemes for multiple choice questions require that
at least one alternative must be true. Otherwise, a division by 0
might appear.  This is an issue, when a subset of alternatives is used
(e.g. 3 out of 10 alternatives). The version know evaluation schemes
with such a requirement and makes sure that in the resulting set, at
least one correct alternative is correct.  Note: this change does not
make sure that the full set of alternatives contains at least one
correct alternative.