Do not force all validation errors to the end of a form.
Caveat: we might miss some validation errors in cases,
where render_item is not used.... but such cases rarely
exist. Provide an log message to identify such cases
that would require further changes. The advantage of
having potentially the message at the right place
is more important for the time being since this is
a usability issue for large forms.