tdom sometimes html-escapes unknown tags
Consider code like
#
#     dom parse -simple -html $form doc
#     ...
#     return [$root asHTML]

Unfortunately, this causes that some tags unknown to tdom
(like <adp:icon>) are converted to escaped tags (&lt; ...).
This can be regarded as a bug. To avoid this problem, we
substitute here the adp_tags in advance. This needs more
investigation in other cases.... The potential harm in this
cases here is very little, but probably, there are other
cases as well where this might harm.