New method "marshall_all_to_file"
This method is similar to "marshall_all", but exports the objects
directly to a file. This can save memory when exporting a large
collection of objects, since the plain "marshall_all" appends to
string, which can get very large, especially due to Tcl's
"double the size when space is needed" policy during "append"
operations.