Drop "ON DELETE CASCADE" on page_template, since this might lead to crippled entries in the database.
When someone delets via operations in the database a form, protentially existing
form pages refering to this item might be left in a crippled state.
The cascade operation deletes the revision (xowiki_page_instance), but leaves
the half-crippled item. The problem does not exist over the web interface,
since it does not allow such items to be deleted while form pages exist.