Make page_template foreign key deferrable
This solves the problem of deleting implicitly child-items, which
depend on each other when deleting an item. When deleting an item, the
deletion of child-items in the content repository happens in an
arbitrary order, which could raise fk constraints. Since the deletion
happens in a transaction, the deferring checking of the constraints
helps, since then the constraint violating item is also already
deleted.

There is no update script necessary, since reload alters the
constraint.