Preliminary cleanup
In this test we will loop through instances of ::xowiki::Page currently in memory to perform some checks. When multiple tests run in the same request, we cannot guarantee that other tests won't create ::xowiki::Page instances that won't comply with the conditions here. We cleanup any existing object in advance to start from a clean slate.