Fixed serious bug killing at least short-text questions in inclass exam
The bug was introduced in [1], by testing for the existence of the
disabled attribute, and when it exists, it was omitting values
reading. The problem is that when form-fields are reset, the
"disabled" attribute is set to 0, leading the exists check to
succeed. In essence, This change sets now the default value of the
form-field to "0", such that it is safe to test it everywhere.
Originally, it was not set by default to save resources (memory and
processing power), but this requires a more careful analysis when
changes happen.

[1] https://fisheye.openacs.org/browse/OpenACS/openacs-4/packages/xowiki/tcl/xowiki-www-procs.tcl?r1=1.368.2.125&r2=1.368.2.126