Added support for anonymous users in chat class
Previous version of the chat class required a login for participating
in the chat. The new version also supports a mixed mode, with
authenticated and non-authenticated users can participate. For
unauthenticated users, the disambiguation is performed over the peer
IP address, similar to the xotcl-request-monitor. Instead of the raw
IP address, an artificial username is displayed for such users (like
e.g. userXXX).

For unauthenticated users, some differences apply:
- one cannot provide a link to the private page
- unauthenticated users cannot access the user-portrait
  of authenticated users.

The GUI changes are only made on chat-skins/chat-classic.js, not on
the other skins.