Extend test suite
when testing FormPage validation, make sure two distinct behaviors are respected:

 1) Rejected values that were part of the request are kept into their original form fields so the user can rework them and resubmit
 2) Other parts of the page, such as the page title, are NOT influenced by data that faild to validate

Rationale: displaying unvalidated information as part of the response can be interpreted as a page injection.

In current codebase, that the title was changed indicates, that rejected information made its way into the :title object member of the FormPage.
The potential consequences of the FormPage setting unvalidated information into its members depends on a number of factors such as formfield logics, callbacks and proper page quoting (to name a few).