Introduce server-side validation for HTML5 date and time formfields
A "formats" parameter can be specified on the formfields indicating one or more formats that we want to enforce. The syntax for any of such format is that of the Tcl clock command.
Default values have been set according to the expected behavior of each form field type.
Empty values are always considered valid. If a field is required, this will be enforced in its own validator.