Render richtext as div reform
Some richtext editors, such as TinyMCE in inline mode, do not expect
to enhance a textarea, but a div.

We could already override the rendering in the richtext editor
subclass, but this has the consequence that inheritance would be
interrupted. This prevents from plugging behavior in subclasses of the
richtext formfield.

This reform refactors render_richtext_as_div from the richtext
formfield into render_as_div on the textarea class. It also supports a
flag for subclasses to instruct the superclass that they should render
the field as div AFTER the whole inheritance chain was applied.