Always use "--" in "dom parse" when document is interpolated
This is a safety measure to make sure that the document parsed is
never confused with an option, when the document starts with a "-". In
the best case, the error message provided by "dom parse" might be
misleading. This might be a problem for user contributed documents
(passed as variables, or return values from functions).

The double dash is supported in tdom since version 0.9.0.